Group plugin: make Allocation a top-level property
Previously, the AllocationMethod was a hidden property determined by the Flavor.
This made it more complicated to detect configuration changes in instances for
the purposes of updates, and would create a hurdle for any future efforts at
supporting composition of Flavors.

To support the schema change in the default Group plugin configuration, the
Flavor SPI is modified to maintain necessary access to a Group's
AllocationMethod.

The punchline of this change is in `group/integration_test.go`:
```
-	// TODO(wfarner): Update this test when addressing the TODO in Spec.InstanceHash().
+	require.Equal(t, "Performs a rolling update on 3 instances", desc)
-	require.Equal(t, "Noop", desc)
```
The test case now validates that a Flavor configuration change results in a
rolling update.

Signed-off-by: Bill Farner <bill@docker.com>

asdf