Adding configurable hostnames to the SoftLayer Terraform plugin.
Currently, the `softlayer_virtual_guest` `hostname` value is ignored and there is no support to define the hostname. This commit adds a new `@hostname_prefix` property; the unique infrakit timestamp is amended as a suffix. For example, if the `@hostname_prefix` is set to `given-hostname`, then the final value is `given-hostname-01234`. If the `@hostname_prefix` value is not set, then the current behavior is not altered.

Signed-off-by: Jacob Frericks <jjfreric@us.ibm.com>