Add basic libvirt instance plugin
Supports basic VM creation but does not yet include storage or networking.
Hardcodes a bunch of things which should likely be exposed as properties, such
as the use of KVM and a serial console.

By default it targets `qemu:///session` which is a per-user session. To drive
the system instance run with `--uri qemu:///system`, this may require
authentication which is not yet supported (dependening on your desktop
environment it might result in a auth popup for each action).

The `Label` method is untested because I couldn't figure out how to cause it to
be called.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>