libvirt: Support metadata using the LinuxKit metadata format.
This uses the same mechanism/format as used by infrakit-instance-hyperkit, that
is, it creates a ISO file containing the spec.Init in a file named `config` at
the top level and attaches that as a CDROM to the device.

The iso image is placed in a configurable libvirt storage pool and we make an
effort to remove it when destroying the instance.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>