Terraform plugin fails to describe instance by tag
The terraform instance plugin's `DescribeInstances` function may fail
if there are multiple groups. The resources types are parsed from the
.tf.json files and stored in a map key'd by the resource type. During
the tag match filtering, if a tag does not match then (currently) no
other resources of the same type are processed.

This is exposed when there are multiple groups since the group name is
included in a tag.

Closes #493

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>