Fix template multipass for group commit
The json passed to the group commit goes through the templating flow
when it is first loaded. This first pass needs to have the "multipass"
support so that additional templating can be completed downstream by
other plugins when additional data is available.

Also, when multipass is enabled, the escape sequence to render the
original var is incorrect. It needs to be a back-tic vs. an escaped
quote.

Closes #515

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>