Expose network device's ip addresses in DescribeInstances.
Thus tries the qemu guest agent first (if it is running) and falls back to
interrogating the libvirt leases database (which is not available with
`qemu://session`).

Filters devices which have no addresses on them.

Modify the example configuration to add a network interface and the necessary
stanza to setup the virtio channel which is used by the guest agent.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>