Append LogicalID to Softlayer instance hostname
Currently, the hostname for the Softlayer terraform resource is either the instance
name (ie, "instance-12345") if the `@hostname_prefix` property is not set; else it's
formatted as `<@hostname_prefix>-<timestamp>` (ie, "given-prefix-12345).

This approach makes sense for workers but having a more predictable naming scheme for
managers would be more usable. This commit updates the naming scheme to use the
`LogicalID` (if set) instead of the timestamp when generating the hostname.

See issue #555.

Note that the instance name remains unchanged; it's only the `hostname` property that
is being updated.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>