Check for quorom group instance changes during PlanUpdate
When doing a `PlanUpdate` on a quorum group, a rolling update is always executed
if the `LogicalIDs` have not changed; this causes an undesired rolling update on a
group of managers.

However, on a scaler group, the hash of the instance definition is also checked to
determine if a rolling update needs to be performed. Therefore, on a scaler group,
no updates are done if the size and instance definition has not changed.

The fix is to add a similar hash check in the quorum function so that a rolling
update is only done if the instance definition changes.

This commit also adds the missing UT to verify the quorum and scaler groups.

Closes #563

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>