Populate "infrakit.attach" tag in tf instance plugin
The filename(s) of non-default `@scope` resources are set as a comma-separated
list of string in the `infrakit.attach` tag.

The challenge here is that Softlayer does not support `,` as a tag character. To
deal with this, the value will be space-separated on Softlayer. The next commit
will handle the parsing out of the tags (so that the `DescribeInstance` SPI returns
comma-separated).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>