Refactor tag parsing flow and handle "infrakit.attach" for Softlayer
The existing `terraformTags` generically parsed tags based on the value
of the `tags` property. Now that we need to change delimiters for the
`infrakit.attach` tag on Softlayer only we need to update this function
to use the `TResourceType` value.

Also, there is no need to pass the `tags` key since the code is platform-
specific.

This commit finishes the work for #561 to write out the `.tf.json` files
and track the related resources in the `infrakit.attach` tag.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>