Terraform plugin locking updates
Terraform uses the filesystem for everything and we need to protect access whenever
we need to modify one or more files. Currently, during Provision, the lock is acquired
to determine a unique filename and then immediately released; this in case the file has
not yet beeen written so a parallel API call could run in the same second and get the
same filename.

This commit protects (via a common lock) the following operations:
* Provision: Hold until the file is written out (the apply is done in a goroutine which
  also gets the lock -- sleep/retry if it cannot)
* Label: Just like Provision, hold until the file is written out
* Destroy: Just like Provision, hold until the file is removed
* Describe: Hold when reading all of the files and running "terraform show"

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>