Destroy related resources on tf instance Destroy
When a tf instance is destroyed, the related resources will also be destroyed
if no other instances reference them.

Note that an instance SPI update is needed to pass the context so that the instance
knows if a rolling update is being down (and the related resources should NOT be
removed) vs a scale down or destroy (and the related resources should be removed).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>