Exit terraform apply polling loop on manager leadership change
During the terraform apply polling loop we will attempt to get the current
manager plugin and check if it is the leader. We will exit the polling loop
if there exists a manager plugin and it is not the leader.

This can be disabled by setting a new `--standalone` CLI arg on the tf instance
plugin; setting this means that the polling loop will never be disabled.

Closes #556

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>