Fix racey maxlife "TestEnsureMaxlife" UT
The problem is that the `DescribeInstances` function is expected to return an
ordered list (based on the instance ID) and the first instance should have an
`Instance.ID` of `0`; this is the instance that is expected to be destroyed.

The problem is order is not guaranteed in the `all` `map`. The fix to first sort
the keys in the `map` and then returned on ordered list.

Closes #614

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>