Import current instance into terraform on startup
Adds support for the following optional CLI params:
`--bootstrap-group-spec-url`: Group spec to use for the imported instance
`--bootstrap-instance-id`: Current instance ID

If these 2 options are set then (on plugin startup):
1. Parse the group spec
2. Decode the instance definition from the group spec
3. Import the current VM into terraform
4. Parse the output of `terraform show` to get the current attributes
   of the current VM
5. Create a `.tf.json` file using the properties in the instance spec
   with the real attributes of the current VM; ensuring that the correct
   group tags (`infrakit.group=<grp-id>` and `infrakit.config_sha=bootstrap`)
   are set
6. `terraform apply`

Note that this only occurs of the current VM is not already in the terraform
state file.

The plugin startup process is blocked until this has completed and will fail
to start if there is an error.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>