TF plugin not applying on manager failover
When a manager leadership changes, Infrakit executes a group `commit` against the
current groups on the new leader. If the groups are up to date then no instance
`Provision` is invoked and the terraform apply goroutine is never started.

This scenario occurs when the current leader manager node is removed. Since the
`.tf.json` file exists for this node, a `terraform apply` will re-create the
node. In order for this to occur, the apply goroutine needs to execute on all
manager nodes.

Note that the goroutine will no-op on non-leader nodes so it can safely
continuously poll on all manager nodes.

Closes #643

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>