Add support to prune tf.json files that are not in tf state
When a resource is destroyed outside of Infrakit/terraform, we need to
remove the associated tf.json file. In order to handle this, the following
logic is added:

On `Provision` the file created has a `tf.json.new` suffix; `terraform apply`
will ignore these files but the Instance's Destroy/Describe/Label functions
will not. The `apply` goroutine will execute:

1. Acquire file lock
2. `terraform refresh`
3. Detect orphan `tf.json` files and delete; we can run `terraform state list`
   to get the list of resource in the state file and then remove any `tf.json`
   files that correspond to resources that are not listed
4. Rename `tf.json.new` files to `tf.json` so that terraform will create them
5. Release file lock
6. `terraform apply`

But naming new files with the `.tf.json.new` suffix we can differentiate between
orphaned resources and those queued up for creation.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>