File and resource name updates for dedicated/global resources
In order to map a given instance spec to existing dedicated resources we need
a consistent naming scheme for the terraform files. Also, in order to differentiate
between different types of dedicated resources, we need a dedicated scope identifier.

This commit adds support to parse an `@scope` identifier for dedicated resources, in
the form of `@dedicated-<scopeID>` (`scopeID` defaults to `default` if not supplied).

The file and resource names for dedicated resources is now:
- Filename: `<scopeID>_dedicated_<instance-XXXX|logicalID>`
- Resource: `<scopeID>-<instance-XXXX|logicalID>-<givenResourceName>`

For for globally scoped resources:
- Filename: `<scopeValue>_global`
- Resource: `<scopeValue>-<givenResourceName>`

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>