Re-attach to orphaned dedicated resources
When a rolling update is performed, the dedicated resources are not
removed; this leaves them orphaned since there is no VM that attaches
to them. When the replacment VM is `Provision`'d we need to determine
and use the orphaned dedicated file.

Also updates the terraform pruning so that it only applies to instance
files. We cannot prune dedicated/global files since these are assumed
to be long lived and we do not want to delete them; if the user changes
the spec then it will be applied to the existing resource. Also, unlike
instances, there is no polling mechanism to ensure consistency so we do
not want to remove the dedicated/globally scoped file since there is
currently nothing that would recreate it.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>