Add support for "/self/dedicated/attachId" tf variable
When processing dedicated resources we detect and use an orphaned resource
file; this commit adds supports to get that attach ID via a template
variable.

This requires the re-factoring of the existing function that writes out the
tf.json.new files; previously this func detected the attachment ID and set
it as a tag. Now, this logic is moved to a new `decompose` function that
breaks up the instance spec into multiple files and returns the dedicated
attach ID value which can then be applied to a template.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>