Use numerical index for dedicated resources
Currently, dedicated resources are in a file that contains the first instance
ID that was attached to it (ie, instance-1234). After a rolling update, this
naming convention causes confusion since the timestamp values do not align.

This commit changes the ID of the dedicated resource to a numerical value that
is incremented when necessary.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>