Add plugin `Option` environment parser
Generic parser that handles the `Envs` values for a plugin's `Options`; the
value is a slice of "key=val" strings that can either be a simple key/value
pair or a reference a key in a local JSON file.

For example, a value of:
```
"Options": {
  "Envs": [
    "key1=val1",
    "key2=file:///run/secrets/my-secret:some-key"
  ]
}
```
Parses the the given `/run/secrets/my-secret` file as JSON, gets the `some-key` value,
and set the `key2` value in the slice. The parsed slice is:
```
["key1=val1","key2=<file-value-of-some-key>"]
```

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>