Add support for Option `Envs` parser and use in tf plugin (#694)
* Add plugin `Option` environment parser

Generic parser that handles the `Envs` values for a plugin's `Options`; the
value is a slice of "key=val" strings that can either be a simple key/value
pair or a reference a key in a local JSON file.

For example, a value of:
```
"Options": {
  "Envs": [
    "key1=val1",
    "key2=file:///run/secrets/my-secret:some-key"
  ]
}
```
Parses the the given `/run/secrets/my-secret` file as JSON, gets the `some-key` value,
and set the `key2` value in the slice. The parsed slice is:
```
["key1=val1","key2=<file-value-of-some-key>"]
```

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>

* Add support to use custom env vars in terraform plugin

The new `Options|Envs` plugin Option is processed to create a slice of
env vars to use on the terraform plugin. These env vars are used on all
`terraform` execs.

The idea is that the env vars would have the specify cloud provider
credentials.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>

* Cleanup terraform execs to get the working dir from the plugin

Since the plugin receiver has the working directory we no longer need
to pass it on the function call.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>

* Update tf plugin Option.Envs parsing to use native templating

Instead of using custom logic with file-uri like values we can simply take
advantage of the fact that the plugin configuration file is already
templated.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>