Ingress controller source key selection support
The L4 plugin `Backends` API returns a `[]instance.ID` slice; the values
in this slice are indexed against the members of a group.

This commit adds support to use a go template when getting the key from the
group members. For example, the L4 plugin `Backends` API could return the
registered IP addresses. These IP addresses could then be indexed against
the specific specifc IP address property in the `instance.Description` for
the group memeber.

The new option on the ingress config yml file is `SourceKeySelector`, see
PR #701 for a similar solution in the enrollment controller.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>