Handle panics in tf provider apply goroutine
When the terraform files are on shared NFS we are seeing random panics in
the `handleFiles` function; this function is executing in a goroutine that
should always be running.

Until we can get to the bottom of what is ultimately causing these, this commit
will recover from the `panic` and keep the goroutine alive.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>