Update tf file system lock to mutex
The current file system lock was intended to prevent multiple
instances of the instance plugin from reading/writing the file
system and it did not serialize operations for the current
instance. The file lock is not necessary since the group controller
only runs on the leader and the tf `apply` goroutine only runs
`terraform` commands on the manager. Moreover, if the leader dies (or
looses network connectivity) while it has the process lock then
we are deadlocked.

This causes issues if multiple `Provision` requests come in (and
quorum groups do not serialize `Provision` requests) because the same
`instance-<timestamp>` files may be used (then then the last request
overrides the file).

The fix is to use a mutex instead of a file lock.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>