Handle files with future timestamps
When looking for terraform file deltas we want to wait until there
has been no updates within the last 3 seconds.

The logic gets complicated if there are file timestamps in the future
and we want to avoid waiting for the local time to catch up to a
future timestamp.

The fix for this has 2 parts:
- Ignore any file timestamp that is more then 30 seconds in the future
- Only wait for no file deltas for 30 seconds (ie, 30 iterations of a
loop with a 1 second wait) vs. an infinite loop that is waiting for
no deltas to be detected.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>