Update loging level for ingress and enrollment controllers
The default `Info` level messages are very noisey. This commit updates most of
the message to `Debug`.

A lot of the messages contain the delta information and, where there actually
are deltas, these are useful at an `Info` level; logic was added to log them
at `Info` when there is something useful.

Also changes the log level to `Debug` for:
- FSM `set.go` Signal code
- Swarm ingress controller `dockerClient` info

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>