Handle manager failover during terraform provision (#791)
* Update IBM Cloud client to get VMs

Updates the generic client to get all VMs in the account.

The `terraform` softlayer code consumes this generic function to get a single
VM that matches the given tags. This will be used in a future commit to import
backend VMs that are missing for the terraform state file.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>

* Break terrafrom import into function

This common function will also be used when importing resources
that are found in the backend.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>

* Handle manager failover during provision

An orphaned resource file can exist for 2 different reasons:
1. Resource has been removed out-of-band from Infrakit, in this case
the resource in the terraform state file will be removed during a
`terraform refresh`.
2. Infrakit failed over to a different manager during a `terraform apply`,
in this case the state file is not updated because the apply never finished.

In order to differentiate between these 2 cases we need to get a list of
existing resources before we execute `terraform refresh`.

When we detect a orphan, we can simply remove the file if the resource
previously existed and is not removed. However, if the resource did not
exist, then:

1. Query the backend cloud for VM with the unique tags
2. If a VM is found then import
3. If no VM is found then prune

The downside of this solution is that is requires platform-specific updates
to perform the backend query; this commit contains the implementation for
IBM Cloud only.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>