Update terraform provider file locking
Currently, the file system lock (shared with the `Provision`, `Destroy`, and
`Describe` functions) is held during the `terraform apply` goroutine while
the resources are refreshed.

We do not need to hold it during this option, since the refresh does not
need to be protected from the `Provision` (since it writes `.tf.json.new` files)
and from the `Describe` (since that is a read-only operation).

This update reduces the number of group `Describe` timeouts since, currently, the
`Describe` is blocked during the `terraform refresh` (which can take several seconds
for larger swarms).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>