Return error when terraform show fails
If `terraform show` fails we need to return on error on the API, if not
then consumers (like the enrollment and ingress controllers) have no
idea the expected data was not returned. This is particularly problematic
for the NFS enrollment controller, when the `id` property is not returned
then it assumes that all instances should be removed (which de-authorizes
all managers from NFS).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>