Handle directory issues in the terraform provider
Currently, if the directory that contains the terraform files either does
not exist or the process does not have read permissions (which can happen
if there are NFS auth issues) then the APIs that read the filesystem do
not error out. When the group controller polls the instance plugin then 0
instances are returned (vs. an error) and the group controller attempts to
provision more.

When the code walks the file system we need to differentiate between when
the file no longer exists (which we need to handle and ignore) and all other
errors (which need to be returned).

This updates the functions to detect these issues and error out
appropriately.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>