Support MissingKey template option
By default, if a template map key is missing, the template rendering
yields a String with `<no value>`. This commit adds support pass
the template `MissingKey` option, see:
https://golang.org/pkg/text/template/#Template.Option

By default, the behavior will not change; this is opt-in.

This commit also updates the enrollment and ingress controllers to use
this new Option so that an `error` is raised.

Note that this new Option only applies to templates that use a Map,
it does not apply to variables in a template (ie, using `{{ var ... }}`).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>