Avoid Deadlock on Failed Commands
When StartWithHandlers() is called, the waitgroup gets incremented as
part of the setup for stdout/stderr handlers. If the command fails to
start, StartWithHandlers() returns prior to calling the handlers. This
leaves the waitgroup in a permanently incremented state. If wg.Wait() is
called, this will result in deadlock. To avoid this, we can call the
handlers prior to returning the error.

This issue surfaced in the terraform provider when we lost access to
NFS on manager nodes. The terraform commands that depended on that
directory failed to start, and the 'terraform apply' routine ended up
in deadlock. This prevented infrakit from recovering once NFS access had
been restored.

Signed-off-by: Dave Freitag <dcfreita@us.ibm.com>