Never remove enrolled instances on parse failure
Currently, when the enrollment controller determines the source key value,
it absorbs all parsing errors. This causes any instance that was previously
enrolled to ALWAYS be removed if it fails to parse.

Ideally, the instance plugin that is returning the group members should always
have the necessary properties (or return an error). However, since the instance
plugin does not know what properties the enrollment controller is looking for,
it does not know what should be returned.

This commit makes the enrollemnt controller more forgiving when instances are
processed that may not have all properties; the trade off is that nothing will
ever be removed until all instances are parse-able.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>