Enrollment Options for parsing errors
The enrollment controller can encounter parsing errors in either the
source instances or the enrolled instances.

A source instance failing to parse only affects the removal of currently
enrolled instances (since we cannot determine the unique identifier of the
source instance due to the parsing error and, hence, we cannot determine if
the instance is currently enrolled). When a source fails to parse there are
two options:
Enable or disable the removal of currently enrolled instances.

If removal is enabled, then the source instance (that failed to parse) will
be removed from the group (it is is currently enrolled).

When a enrolled instance fails to parse, this only affects the addition
of source instances and there are two options:
Enable or disable the addition of source instances.

If addition is enabled, the enrolled instance (that failed to parse) will
be added again (if it is in the source instances).

This commit adds these `Options` to the enrollment controller configuration,
the default is that removals and additions are ENABLED on parse errors.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>