Update default Option parsing for the source error ops
When the enroller is created, default values may be set for the
`SourceParseErrOp` and `EnrollmentParseErrOp` values. These values
will override the hardcoded default values (where both destroy and
provision are enabled).

When the spec is committed or updated, these values can be overriden.
The verification of the values is moved to a utility function so that
it can be invoked both when the enroller is created and when the spec
is updated.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>