Cache terrform instances
This commit caches the output to the `DescribeInstances` command. The cache
needs to be invalidated whenever one of the following operations is done:
- Provision
- Label
- Destroy

Also, after `terraform refresh` is executed then the properties may have have
changes OR input `tf.json` files need to be pruned; therefore, the cache also
needs to be invalidated after each refresh.

This commit also updates the locking to use a read/write lock so that multiple
describe operations can run in parallel.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>