Add tests for IBM Cloud clients
Creates an `interface` for the backend API calls so that we can
test the volume auth plugin and the terraform plugin's use of the
client.

Creates a "fake" Softlayer client so that the data returned can be
specified and the arguments tracked.

Lastly, this fixes the error handlings in the actual code (returning
`nil` vs the error) that the new UT code detected.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>