Swarm flavor group and instance plugins
Swarm instance plugin: Implements `DescribeInstances` and `Destroy` only
- `DescribeInstances` returns all swarm nodes (no filtering) with
the swarm node ID as the instance `ID`, the "link-label" as the
`LogicalID`, all engine labels as the `Tags` (plus a `name` tag added
with the node `hostname`), and the entire node struct as the `Properties`.
- `Destroy` still needs to be implemented.

Swarm group plugin: Implements `DescribeGroup` and `Size` only
- `DescribeGroup` retrieves all swarm nodes from the instance plugin and
excludes any nodes without a `LogicalID` (those that do no have a "link-label")
and and that are not in a `ready` state.
- `Size` returns the number of nodes with the same filtering as `DescribeGroup`.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>