Expose Destroy error during rolling updates
Currently, when the instance `Destroy` fails, the error does not stop
the loop that is waiting for the group is quiese. The loop then will
never terminate because no replacment instance will ever be `Provision`'d
because the `Destroy` failed.

The fix is to expose the `Destroy` error so that the loop is terminated.

This also updates the UT flow to use custom tag values to denote when
the test plugin should return an error.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>