Use instance hash to determine rolling update members
The group hash is commited at the beginning of the update cycle, it it not
guaranteed to reflect that actual instance hash of all members. During an
update we cannot rely on the group hash to determine if any instances need
to be updated; we need to check the instance hash on each member.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>