Update swarm/manager Flavor drain to be idempotent (#851)
When a manager node is removed, the following occurs:
- The associated flavor `Drain` function is invoked, this involves
a node demotion and then a swarm node leave or a swarm node remove
- The instance plugin's `Destroy` function is invoked

If an error occurs in the processing then the instance may not be
removed from the group. Currently, depending on the error, all future
group operations that invoke a `Drain` will always fail because
the `Drain` will error out if the node is not still a manager.

The fix is to make the `Drain` idempotent so that it return `nil` if
the node has already been removed from the swarm. If the node is still
in the swarm as a worker then it will be removed from the swarm.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>