Handle stale swarm version in swarm store
On startup the manager plugin attempts to write to the backend
store. In swarm mode this can fail with:
Error response from daemon: rpc error: code = 2 desc = update out of sequence

This commit detects this error, retrieves an updated version, and
attempts it again (up to 10 times with an increasing wait time).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>