Update enrolling logging
We log an `Info` message with the full `Instance.Description` value whenever
an instance is added or removed; this gets really noisey for VMs since the full
userdata value (with the full boot script) is logged.

This commit updates this flow so that the instance `ID` is logged or, if the
`LogicalID` is set, then the instance `ID:LogicalID` is logged in the `Info`
message.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>