Change rolling update quiesced loop to handle Unhealthy nodes
Now that we have definition of a "Healthy" node has been enhanced to
require a node in `ready` state (and `reachable` for managers) we need
to be able to handle unhealthy nodes. Currently, the updating loop
simple errors out when it finds an unhealthy node. This commit changes
the logic to allow them and then re-check on the next poll.

Note that, in the future, we may want to add an maximum number of
times that a node is unhealthy before we give up since, with this change,
the update will continue indefinitely if there exits a single unhealthy
node.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>