Clear cached tf instances when not running on leader
If the tf plugin is not running on the leader then the cached instances
may not be valid. We need to clear the cache whenever we detect that
we are not running on the leader.

Also fixes the locking when clearning the cache after the initial
apply (the write lock is needed when clearing).

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>