Pass "pretend" flag on plugin creation
The "apply" go routine is always started in the test code because we
are not setting the "pretend" value on the plugin struct until after
it is created; this is causing race conditions in the tests.

This commit moves the "pretend" setting to the construction of the plugin
so that we can short circuit the creation of the "apply" go routine.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>