Fix racey Manager "TestStartOneLeader" test
This was introduced in https://github.com/docker/infrakit/pull/856

Now that a 2nd `CommitGroup` is being executed on leadership, we need
to synchronize after the 2nd API call. Note that this work was completed
for the other test in this package, `TestChangeLeadership`, and the same
logic needs to be applied to the `TestStartOneLeader` test.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>