Remove config.hash=bootstrap tag on imported tf VM
We want to keep the import flow generic. Currently, the tf plugin
assumes that the VM being imported is the initial manager and it
sets the `infrakit.config.hash=bootstrap` tag. Instead of doing this
in the plugin, the initial VM should simply be created with this tag;
this is consistent with how the `infrakit-link=bootstrap` tag needs
to be set.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>