Import updates for terraform v0.10+
Starting in teraform v0.10, the `tf.json` file needs to be created
before the resource is imported.

The current import flow does not have the file data at import time,
we rely on running `terraform show` on the imported resource in order
to create the actual `tf.json` file. In order to satisfy the terraform
import requirement, we create a dummy `tf.json` file that has the minimum
content that terraform requires; this file is then removed after the
import has completed.

The normal import flow will then parse the `terraform show` output and create
the actual `tf.json` file for the resource.

Signed-off-by: Steven Kaufer <kaufer@us.ibm.com>