Fix fsnotify and apimachinery dependency errors
This commit pins the hpcloud/tail dependency to the fix in the PR branch
for hpcloud/tail#138 which addresses problems due to the fsnotify github
org name changing from go-fsnotify to fsnotify
(see go-fsnotify/fsnotify#1).

This change also effectively bumps the vendored hpcloud/tail version to
`1.4.7`, no impact to tests has been observed.

When this upstream PR is merged we can point again at the hpcloud/tail
repo.

This commit also addresses an issue introduced when migrating the
dependencies from Rancher Trash to dep, the `dep init` used to
populate the `Gopkg.toml` file had interpreted the pinned value for
k8s.io/apimachinery as a revision and not the branch which it pointed
to. This then resulted in errors when trying to add new dependencies
or freshly `dep ensure` the current locked versions as it was treating
`release-1.7` as a revision.

The fix involves pinning k8s.io/apimachinery to the `release-1.7`
branch again in an override.

It should be possible to remove this when moving the k8s.io/kubernetes
dependency to a version `> 1.7.5` (there is no corresponding `1.7.5`
tag on the apimachinery project, the earliest being `1.7.6`).

Signed-off-by: John McCabe <john@johnmccabe.net>